helpers for extracting exception data from clojure.lang.IExceptionInfo
In recent improvements to Datomic, many exceptions will be thrown as either IllegalArgumentException or IllegalStateException. The concrete types are actually subtypes of these standard Java exceptions, which also implement Clojure’s IExceptionInfo interface. These helpers are there to match on this type, and to help extract the Clojure map that holds the exception data.